+++ src/DotNetStac/StacItem.cs
@@ -1,22 +1,19 @@
-using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Runtime.Serialization;
-using Stac;
 using Stac.Converters;
 using GeoJSON.Net.Geometry;
 using Newtonsoft.Json;
-using Stac.Extensions;
 using System.Linq;
 using System.Net.Mime;
-using Newtonsoft.Json.Linq;
 using Semver;
 using System.Collections.Specialized;
 using System.IO;
+using System;
 
 namespace Stac
 {
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
+    // [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class StacItem : GeoJSON.Net.Feature.Feature, IStacObject
     {
         public const string MEDIATYPE = "application/geo+json";
@@ -29,18 +26,21 @@ namespace Stac
                         base(geometry, properties, id)
         {
             Preconditions.CheckNotNull(id, "id");
-            StacExtensions = new Collection<string>();
+            StacExtensions = new SortedSet<string>();
+            Root = new StacItemRootPropertyContainer(this);
             StacVersion = Versions.StacVersionList.Current;
             Links = new ObservableCollection<StacLink>();
             (Links as ObservableCollection<StacLink>).CollectionChanged += LinksCollectionChanged;
             Assets = new Dictionary<string, StacAsset>();
+
         }
 
         public StacItem(StacItem stacItem) : base(Preconditions.CheckNotNull(stacItem, "stacItem").Geometry,
                                                   new Dictionary<string, object>(Preconditions.CheckNotNull(stacItem).Properties),
                                                   Preconditions.CheckNotNull(stacItem, "id").Id)
         {
-            this.StacExtensions = stacItem.StacExtensions;
+            this.StacExtensions = new SortedSet<string>(stacItem.StacExtensions);
+            this.Root = new StacItemRootPropertyContainer(this);
             this.StacVersion = stacItem.StacVersion;
             this.Links = new ObservableCollection<StacLink>(stacItem.Links);
             (Links as ObservableCollection<StacLink>).CollectionChanged += LinksCollectionChanged;
@@ -78,7 +78,7 @@ namespace Stac
         /// The STAC version the Item implements
         /// </summary>
         /// <value></value>
-        [JsonProperty("stac_version")]
+        [JsonProperty("stac_version", Order = -10)]
         [JsonConverter(typeof(SemVersionConverter))]
         public SemVersion StacVersion { get; set; }
 
@@ -86,7 +86,7 @@ namespace Stac
         /// A list of extension identifiers the Item implements
         /// </summary>
         /// <value></value>
-        [JsonProperty("stac_extensions")]
+        [JsonProperty("stac_extensions", Order = -9)]
         public ICollection<string> StacExtensions { get; private set; }
 
         /// <summary>
@@ -94,7 +94,7 @@ namespace Stac
         /// </summary>
         /// <value></value>
         [JsonConverter(typeof(CollectionConverter<StacLink>))]
-        [JsonProperty("links")]
+        [JsonProperty("links", Order = 5)]
         public ICollection<StacLink> Links
         {
             get; private set;
@@ -105,15 +105,30 @@ namespace Stac
 
         # endregion IStacObject
 
-        [JsonProperty("assets")]
+        [JsonProperty("assets", Order = 4)]
         public IDictionary<string, StacAsset> Assets { get; private set; }
 
         /// <summary>
         /// The id of the STAC Collection this Item references to
         /// </summary>
         /// <value>gets the collection id</value>
-        [JsonProperty("collection")]
-        public string Collection { get; internal set; }
+        public string Collection
+        {
+            get => Root.GetProperty<string>("collection"); 
+            set
+            {
+                if (value != null) Root.SetProperty("collection", value);
+            }
+        }
+
+        /// <summary>
+        /// Item root extended data
+        /// </summary>
+        /// <value></value>
+        [JsonExtensionData]
+        public IDictionary<string, object> RootProperties { get => Root.Properties; internal set => Root.Properties = value; }
+
+        private StacItemRootPropertyContainer Root;
 
         [OnDeserialized]
         internal void OnDeserializedMethod(StreamingContext context)
@@ -126,6 +141,7 @@ namespace Stac
             {
                 asset.ParentStacObject = this;
             }
+            StacExtensions = new SortedSet<string>(StacExtensions);
         }
 
         [OnSerializing]
@@ -133,6 +149,7 @@ namespace Stac
         {
             if (BoundingBoxes == null)
                 BoundingBoxes = this.GetBoundingBoxFromGeometryExtent();
+
         }
 
         public bool ShouldSerializeStacExtensions()
@@ -143,5 +160,10 @@ namespace Stac
 
         [JsonIgnore]
         public IStacObject StacObjectContainer => this;
+
+        public object RasterExtension()
+        {
+            throw new NotImplementedException();
+        }
     }
 }