+++ src/DotNetStac/StacObjectLink.cs
@@ -53,5 +53,7 @@ namespace Stac
             set;
         }
 
+        [JsonIgnore]
+        public IStacObject StacObject => stacObject;
     }
 }